pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[tool.pytest.ini_options]
addopts = "-q --maxfail=1 --disable-warnings"

[project]
name = "qa-framework"
version = "0.1.0"
dependencies = [
  "pytest>=8.3",
  "pytest-xdist>=3.6",
  "pytest-parallel>=0.1.1",
  "pytest-playwright>=0.5.0",
  "playwright>=1.48.0",
  "pytest-html>=4.1.1",
  "pytest-rerunfailures>=14.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.2",
  "pydantic>=2.9",
  "Django>=5.1",
  "django-environ>=0.11.2",
  "structlog>=24.1.0",
  "allure-pytest>=2.13.5"
]
